feat: consolidate utility imports for better organization
