+++ src/constants/api.constants.ts
@@ -1,4 +1,4 @@
-import { getBaseURL, getWebUIURL } from "@utilities/url.utilities";
+import { getBaseURL, getWebUIURL } from "@utilities";
 
 export const BASE_URL = getBaseURL();
 export const WEB_UI_URL = getWebUIURL();
