+++ src/utilities/index.ts
@@ -17,3 +17,4 @@ export {
 } from "@utilities/fileSystem.utils";
 export { omit } from "@utilities/omit.utils";
 export { errorMessageWithLog } from "@utilities/messages.utils";
+export { getBaseURL, getWebUIURL, getOrganizationId } from "@utilities/url.utilities";
